ci-operator/step-registry/openshift/credentials-request-freeze: Only hash on no-diff
The hashing landed in e11b103a71
(ci-operator/config/openshift/release: Add credentials-request-freeze
for 4.7 and 4.8, 2021-12-01, #24166) with the following commit message
motivation:

  I'm also adding a sha256sum command to make it easier to figure out
  what's going on when the diff gives surprising results like
  unexpectedly matching beteween 4.6.1 and a recent 4.6 nightly [4].
  ...
  [4]: https://github.com/openshift/release/pull/24166#issuecomment-984068893

The hashes are pretty chatty, though, and we don't want them
distracting from the diff when we do actually notice a diff.  This
commit shifts them below the diff, so we will no longer hash when
there is a diff (helping readers focus on that diff), but we will
still hash when there wasn't a diff (helping readers debug when a
lack-of-diff was surprising).