release-controller: Add `ocp-4.11-rhel85.repo`
For RHCOS, we must have the ability to choose (as a pull request)
which version of RHEL we use to build.

Basically, I want to make https://github.com/openshift/os the canonical
source for this.

In preparation for that, add a split off rhel85 repository.

I think we should also add a corresponding rhel86 version, but
that can come in a bit.