ASH: Check for techpreview CRs
Add check if CredentialRequest has TechPreview

In 4.10 there was an additional CR for CAPI. This
CR is available in the release image and applied
directly to the cluster during install.
The CAPI CR references the openshift-cluster-api
namespace which doesn't exist unless the techpreview
flag is enabled causing the cluster to not bootstrap.

This PR checks if the flag is annotated and skips
those CRs.