Remove `release-4.1` merge criteria
Because of https://github.com/kubernetes/test-infra/issues/24089, the
merge criteria for `release-4.1` apply on `release-4.10` too, which
normally would not be a problem, because `release-4.1` criteria are
strict and require various approvals which non-GA PRs are not
supposed to have. Unfortunately some 4.10 PRs *did* have them, which let
them merge unexpectedly.

If we remove `release-4.1` from the list of branches included in the
queries that enforce post-GA merge criteria, we are safe from
interference. We are keeping `release-4.1` in the "complement" queries,
so the relaxed complement queries will not apply on actual `release-4.1`
PRs either - Tide should be completely disabled for the actual
`release-4.1` PR.

Created with https://github.com/petr-muller/ci-tools/tree/release-41-pwn
version of `determinize-prow-config`.