Hypershift: Archive guest cluster dump in conformance periodic
Analyzing certain failures requires to look at the dump of a guest
cluster. Doing this through the gcs ui is very difficult, it is more
efficient to do this locally. The sheer number of files in the dump
means that downloading it takes very long (about seven minute) despite
it not being big in size (about 20m). So simply create an archive in
addition to the raw files.