Retire `jenkins-ci-provisioner` secret in Vault
That name is confusing. We do not use Jenkins for several years. The
Vault record was used to store GCP credentials for the
`openshift-gce-devel-ci` account. I made a copy in Vault with the
identical content: new name better indicates the secret's purpose and
is also similar to existing Vault record for the second GCP account, so
now the two secrets are listed next to each other.

The SSH keys from the `jenkins-ci-provisioner` were also used in some
OpenStack secrets, so I made copies of them in the `openstack` record,
so that GCP and OpenStack secrets are not coupled.

For now I kept `jenkins-ci-provisioner` secret around, for the case we
need to revert this PR.