Trigger static-ip job daily instead of 6 times a day
We previously had some problems with this job (and specifically with the
previous generation that tried running 4 tests in the same job). It
seems like we're in a better place in terms of stability of this job,
with most of the failrues looking like general failures that can happen
on all test-infra job.