MGMT-9470: Publish catalog index to quay
The current solution build a bundle that points to the latest tag of the
assisted-service images which makes its content unpredictable. This new
job pins the assisted images in the bundle and publishes a catalog index
created from it.

In order to work around the docker-in-docker situation, the job uses
umoci along with skopeo to build and publish the bundle and the catalog
images.