use cluster pool for cincinnati-graph-data tests
use the cluster pool to reduce the time it takes to test
cincinnati-graph-data

this is needed as the cluster installation takes considerable time
and results in delayed merge for the releases.

Also, the cluster pools dont have a 4.11 cluster as per a previous
job run. so decided to move it to 4.10.