CNO: Add Kuryr dev and QE for backport approval
New backport process suggests adding all the interested parties as
backport approvers in shared projects. This commit allows Kuryr dev team
member to add `backport-risk-assessed` to the CNO PRs. The idea is that
we'll exercise this power for changes that can only affect Kuryr
deployments. Then OpenStack QE team member is allowed to add
`cherry-pick-approved`. It's assumed that they'll use it only to bugs
assigned to them.