Ensure Spyglass is passed .out files for coverage lens
Spyglass expects to receive the raw Go coverprofile and at most one 
additional html file. Typically the coverprofile is called cover.out 
which is not picked up by the previous regex.
When no raw coverprofile is provided, the spyglass lens will load but 
there will be a javascript parse error causing the lens to be blank.

We should ensure that .out files are also included so that users do not 
have to name the coverprofile as a .xml (as it is not XML)