openstack: add support to test DPDK on Mellanox hardware
This will separate the tests between Intel & Mellanox, so both can test
DPDK.

This adds a new workflow for testing DPDK only. We haven't made
SR-IOC working yet with our Mellanox device and we're working on it. In
the meantime, we only need to test DPDK.

The workflow will do the following:
* Create a performanceprofile
* Prepare the workers for vhostuer and noiommu
* Scale the worker to 1
* Run DPDK tests