machine-config-operator: Drop `skip_if_only_changed` for all `optional: true` jobs
The machine-config-operator is a key component of OpenShift 4 that
touches on many things, from OKD to single node to bare metal magic
to networking.  We even used to handle etcd.

We have accumulated a giant pile of jobs due to this.  But
not every job needs to run on every PR - it's at best a waste of money,
and at worst a huge distraction because every pull request is accompanied
by a giant pile of red "X"'s by default.

Now, many of these jobs are actually `optional: true` - their failure
will not block merges.

And in theory, they shouldn't run by default; but many of them got
a `skip_if_only_changed` annotation to avoid them only docs changed,
but that annotation isn't implemented in the way one would expect.
It appears to be a simple *inverse* of `run_if_changed`.  Which
means that all these optional jobs are still triggered for not-documentation
PRs.

So by just dropping this, we should return to not running these jobs
by default.  People touching things that might affect bare metal
specifically will need to e.g. invoke `/test e2e-metal-ipi` explicitly
now, and similar for the other jobs.