WIP Add initial hybrid-cloud-patterns org testing repo
Do not merge this as is.

This is an initial WIP PR to onboard the https://github.com/hybrid-cloud-patterns/ci-testing
repo.

The longer term goal is to be able to leverage prow for a few more tests
(both containerized tests and end-to-end tests) and additional
non-testing repos in our hybrid-cloud-patterns org.