Add GW migration jobs as periodics
local gw to shared gw migration and vice versa have presubmit
jobs that will run on PRs. This commit adds those same jobs as
periodics. This allows for some history of the job running
on release images to build up making it easier to compare for
failures we may see in a pre-submit.

- 4.11+/master jobs will run every 2 days
- 4.10 will run every 3 days
- 4.9 will run every 4 days

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>