release-ocm-2.5 assisted-service postsubmit should publish to ocm-2.5 tag on quay
Sets `OPERATOR_MANIFESTS_TAG_TO_PIN`, `REGISTRY_CATALOG_REPOSITORY_TAG`
and `REGISTRY_BUNDLE_REPOSITORY_TAG` to non-default values for the
`release-ocm-2.5` `operator-catalog-publish-to-quay` job (the defaults
are only suitable for the master branch).

When publishing the Infrastructure Operator bundle and catalog to quay
for the master branch, we use `pin-latest.py` toscan the operator's
manifests (clusterserviceversion in particular) for the "latest" tag and
pin it to a particular image manifest digest. This is done using the
pin-latest.py script and it by default uses the "latest" tag.

`OPERATOR_MANIFESTS_TAG_TO_PIN` makes `pin-latest.py` use a custom tag
for pinning, usually the original tag in the manifest is "latest" but in
release branches the tag is different so we need to inform the script
of what the tag is.

We also set `REGISTRY_CATALOG_REPOSITORY_TAG` and
`REGISTRY_BUNDLE_REPOSITORY_TAG` to publish the 2.5 bundle/catalog to a
different tag than the master bundle/catalog, so the 2.5 `operator-catalog-publish-to-quay`
job doesn't overwrite the images creates by the equivalent master
job.