MGMT-9588: publish assisted-service Python client
We previously have done that in Jenkins, but ever since we removed the
jobs no more Python client versions have been uploaded to PyPI:
openshift/assisted-service#3373

As an example to that previous job that we had:
http://assisted-jenkins.usersys.redhat.com/job/assisted-service/job/master/15713/console

This PR aims to add a postsubmit job that publishes our packag to PyPI
on merges to master branch. Building of the package is already part of
presubmit job verify-generated-code, so except from the actual CD code
we're testing all this flow pre-submit.