Add doc explaining ovn workflow without ovn steps
It was confusing to see a job with ovn workflow not having
any steps specifically configuring OVNKubernetes. If you
dig in to the dev-scripts code you will find that when the
IP_STACK variable is v6 or v4v6 then OVNKubernetes is the
default NetworkType. Having this note here might save
someone some time understanding this in the future.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>