Disable autosize kubelet config Observed that the autosize reservation for m5.4xlarge was: cpu=0.11,memory=5.36Gi Observed that the autosize reservation for c6a.24xlarge was: cpu=0.31,memory=10.44Gi
Our historial m5.4xlarge setting reserved 1 full core for the kubelet. Based on this
historical setting and the fact that timeouts were observed on the 96cpu c6a,
we believe the autosize select sufficient memory but insufficient CPU for our workloads.

This reverts the autosize kubelet configuration option and introduces instead a configuration
sepcifically for 96 core machines.