machine-config-operator/layering: Promote builds into `rhcos-devel`
Part of https://issues.redhat.com/browse/MCO-238

We want to generate a release image containing the latest MCO code
and the latest RHCOS new-format bits, and to do that we need a
build of the latest MCO from the `layering` branch.

I am not entirely sure this is correct, but it seems worth a try.