OpenStack: Ensure there is only one LoadBalancer section
Starting with OpenShift 4.9, we set the `use-octavia` option in the
cloud.conf's LoadBalancer section. This caused the
ipi-conf-openstack-enable-octavia-commands workflow to add a duplicate
LoadBalancer section. This was not problematic since it set the same
option to the same value, however this can quickly become confusing once
we set more options.