Fix Pods readiness check for the sriov-networknodepolicy step
When using the operators-subscribe step an extra Pod in created
in the sriov-network-operator namespace and ends with Completed
state, which breaks the checking if all the containers are ready
for Pods in the namespaces and ends up never creating the node
policy. This commit fixes the issue by checking if all the Pods
are either Running or Completed.