core-services/prow/02_config/generate-boskos: Almost zero AWS-1
It's flooded with IAM resources, which is making for IAM API
contention, which is causing AWS to throttle us.  Stop launching
clusters in this account until we have cleared out the leaked
resources.

I didn't completely elimitate it, because I wanted to keep the pool
name, and presubmits didn't like it zero [1]:

  time="2022-03-25T18:13:52Z" level=fatal msg="Config validation failed" error=".6.max-count: must be >0"

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/27301/pull-ci-openshift-release-master-boskos-config/1507420174137954304