hypershift: test hypershift upgrades using payload images
Before this commit, upgrade tests were exercising upgrades by overriding
CPO directly rather than using payload images. This commit makes it so that
a new payload is built from the PR which is used to run the upgrade using
the normal flow rather than overriding CPO on the hosted cluster.