Allow Windows VMs to be created without a cluster
This commit fixes an issue where the creation of a Windows VM in vSphere
would only be possible if an OpenShift cluster was created. This is due
to the cluster's VM folder being created as part of the install process.

To fix this I have made it so that if the required VM folder does not
exist when attempting to create a Windows VM, the folder is created, and
will be subsequently cleaned up when removing the VMs.