cluster-api-provider-openstack: main unit-test
Add a configuration file for the new 'main' branch of
cluster-api-provider-openstack, which is the actual downstream of the
kubernetes-sigs repository (as opposed to the 'master' branch, which
contains a fork of it).

The 'main' branch does not build into an image directly. Rather, it is
imported as a dependency into openshift/machine-api-provider-openstack.

With this commit, unit tests are required for merging PRs in the 'main'
branch of cluster-api-provider-openstack.