Remove metal from blocking jobs
TRT has unfortunately reached the point where we believe we have to
remove all metal jobs from the list of payload blockers. We know these
jobs provide great coverage and have found real issues in the past, but
sadly in their current state they are preventing us from shipping
payloads which carries an immense cost to the organization.

TRT is discussing with the metal team the requirements to get back in as
a metal job.