Add a role that allows `system:authenticated` to read the `ocp` imagestreams
In https://github.com/openshift/release/pull/27388
I am trying to add a periodic job that creates a custom release image.

While I could create a custom serviceaccount and roles for this,
that feels a bit like overkill.  I can't think of a downside
of allowing anything on the cluster to *read* the `ocp` imagestreams,
so let's just do that.