openstack/bastionproxy: bind dnsmasq on the private network
Before we would try to bind the dnsmasq process on the bastion floating
IP, which has been problematic with the recent move to centos9-stream,
where the dnsmasq process failed to start.

In reality, it's more realistic to bind it to an IP that actually exist
*on* the host, which is already how we did when the job was configured
with `proxy` type. The `byon` type will do the same and use the machine
network port for binding squid & dnsmasq.