assisted-baremetal-setup step should use the CI catalog image
In the openshift/assisted-service repo, The CI builds a catalog image
for us and calls it `ci-index`.

On some jobs, the assisted-baremetal workflow (which uses
assisted-baremetal-setup) is used to install a cluster on which we
immediately install the assisted operator on, using that step's
POST_INSTALL_COMMANDS environment variable. We use that environment
variable to run `make deploy_assisted_operator` command. That command
uses the catalog image specified in INDEX_IMAGE, which by default uses
an image somewhere in quay. We instead would like the operator that gets
installed to be installed from the catalog image created by the CI, so
we'll actually be testing the operator code that's changed and not
something that's periodiocally published to quay.