openstack: remove BYON (replaced by proxy)
We first introduced BYON testing by a simple architecture where we had a
bastion node acting as a proxy for prow, which was connected to a
pre-provisioned network (BYON).

When adding support for testing clusters running behind a proxy, we
re-used the work done for BYON and now we don't need to support the 2
setups. We can just use "proxy" config type every time we need to test
BYON, this will be fine.

In order to avoid complexity in our scripts and testing, let's just keep
"proxy" config type for now and remove "byon".