core-services/prow/02_config/openshift/cincinnati-graph-data/_prowconfig: Bot doesn't need lgtm or approval
We trust it, and it's no longer authorized to add its own labels:

  * FAILED channels/fast-4.7: Promote 4.7.48. It was promoted to the
    feeder fast by 3ea452a650 (Merge pull request #1745 from
    openshift/promote-4.7.48-to-fast, 2022-04-13) 1:39:00.240438
    ago. 403 {"message": "Must have admin rights to Repository.",
    "documentation_url":
    "https://docs.github.com/rest/reference/issues#add-labels-to-an-issue"}

Also drop the backport business for all authors, because we don't have
release branches in this repository.

I'd also like to drop bugzilla/invalid-bug, because we don't tie
changes to bugs.  But CI presubmits didn't like that [1]:

  {"component":"checkconfig","file":"k8s.io/test-infra/prow/cmd/checkconfig/main.go:90","func":"main.reportWarning","level":"warning","msg":"the following orgs or repos enable the bugzilla plugin but do not forbid the bugzilla/invalid-bug label for merging: [repo: openshift/cincinnati-graph-data]","severity":"warning","time":"2022-04-13T18:08:18Z"}

I poked around a bit trying to find a way to drop the Bugzilla plugin
from the graph-data repo, but couldn't find the incantation, so for
now I'm keeing the invalid-bug label in these queries.

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/27826/pull-ci-openshift-release-master-config/1514304177340157952#1:build-log.txt%3A10