Hypershift: Enable slack reporting for 4.10 and rename for consistency
This re-enables the slack reporting for the release-4.10 job and renames
the `main` job to be named consistently. We didn't do that previously
because we didn't want to mess up testgrid, but we already messed up
testgrid because what it currently has in the 4.10 tab is in fact for
4.11 and and what it has for 4.11 is referencing a job that doesn't
exist anymore.