MGMT-10164: add verify-generated-code job for release-ocm-2.5 branch
Somehow we don't have the verify-generated-code job for release-ocm-2.5
branch in assisted-service repository.
It means we do not validate that all generated code has been committed
together with the corresponding "source" of changes (e.g. swagger.yaml
changes, generated mocks, etc.).

This change adds this job, similarly to how it's configured for the
master branch.