openshift/assisted-test-infra repo live-iso-with-worker job
Add an assisted-test-infra job that runs the SNO BIP + Worker flow

This job is being added so the code in openshift/assisted-test-infra#1639
can be tested before merging.

It will only run if files matching the pattern .*bootstrap_in_place.*
ever change in the future

This job only makes sense on OCP 4.11 so that's why the OCP 4.11 image
is explicitly being used.