Hypershift conformance: Use correct image for nested mgt cluster
Currently we use bin to create and destroy the nested mgmt cluster. This
only works if the release-4.10 version of the Hypershift cli is compatible with
the main version deployed in the root management cluster. Because this
in't the case the conformance tests currently fail to create the
management cluster.