assisted-service: promote index image
This PR promotes `ci-index` as assisted-service-operator-index in the
`ocm-2.6` scope so that it can be later consumed from the Hypershift
repository.

This is a prerequisite for running ZTP jobs from other repositories as
those are installing Infrastructure Operator from the index image and
not by directly deploying the service image.