Don't set a default value for `baremetalds-sno-setup-ref`'s `TEST_FUNC` env
Using `test_bootstrap_in_place` as a default value for `TEST_FUNC` is
not a great idea because the name of the Python file containing the bip
tests is `test_bootstrap_in_place.py`, so all tests within it get
matched by the pytest `-k` expression, instead of just the test called
`test_bootstrap_in_place` which was what was intended by this default
value.

Leaving the `TEST_FUNC` variable to be empty is good enough because
older revisions of assisted-test-infra simply only have a single test
in that file, and newer WIP revisions of `assisted-test-infra` which add
more tests to that file also rename the `test_bootstrap_in_place` test
to `test_bootstrap_in_place_sno` and force the Makefile to default
`TEST_FUNC` to `test_bootstrap_in_place_sno`.