release-controller/_repos: point the 8.5 mirror and 8.5 URLs
CI jobs run against `openshift/os` rely on the repo mirror to build
RHCOS for test.

Currently, we are building/testing using RHEL 8.5 content. The repo
defintion here, which has `rhel85` in the name, used URLs that point
to the latest available RHEL 8 release, which is now 8.6.

Change these URLs to point to RHEL 8.5 content, appropriately.