ocp-4.11 repos: align after 8.6 GA
with RHEL 8.6 GA we can now pin to it for the base OS and codeready, and
get rid of beta repos; this aligns with ART config.

NFV and virt aren't configured by ART, so these were left pinned to 8.4
in ocp-4.11-rhel8.repo as they have been, and this was copied to
ocp-4.11-rhel86.repo so they are now identical.