Add label for publishing openshift/dynamic-plugin-sdk
We are looking into the ability to auto-publish our dynamic-plugin-sdk to our NPM repo in order to aid in the CICD flow
of changes. See https://issues.redhat.com/browse/HAC-806

Our goal is to use PROW to do this through a custom label on our PRs that will trigger our test script to:
1. Premerge run an npm publish --dry-run with the openshift bot token
2. Postmerge run the actual publish

We could possibly use Jenkins or CircleCI, however we'd like to try using PROW for this first.