Add explicit metal launch jobs for ovn v4 and v6
This is just overloading existing clusterbot launch
jobs that can already be launched. It could be more
user friendly and clear up some confusion if someone
passes a parameter that might already be the default
and otherwise not accepted.

example:
"launch <release> metal,ovn" will currently deploy
an ipv4 cluster. Passing an ipv4 parameter would be
rejected. With this "metal,ovn,ipv4" would also
work.

similar examples for ovn and v6 or dualstack.
currently, you can't pass ovn to v6 or dualstack
because ovn *has* to be used for those stacks.
With this, passing ovn will now be accepted.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>