Making the e2e test to run against the PR changes instead of nightlies.
The e2e test in DTK is mainly used for running on nightly builds.

On PRs we are running the e2e test as well but on the nigthly payload,
which means that the e2e test is only making sure we are not breaking
the test itself but doesn't actually verify anything related to DTK.

This changes is adding an optional job that will run on the latest
payload with the "new" DTK integrated in it.

The issue with such job is that it may occur that the
machine-os-content will change and get a new RHEL version or a new
kernel in the release image we get in CI but since DTK hasn't changed, the DTK
image won't be rebuilt as part of the job, therefore, the job will fail on
OS/kernel mismatch.

This new job is a best effort to make it work. In addition I am checking
how can we force the DTK image to be rebuilt on every run of the job,
even if there were no changes in the DTK image.

Also updated the jobs names for extra clearness.

Signed-off-by: Yoni Bettan <yonibettan@gmail.com>