Bump timout of ipi upgrade jobs to 5hrs
With the rebase to kubernetes 1.24 the cluster upgrade is taking
longer then is usual. This is causing us to hit the 4hr CI limit.
Temporarily bump the limit up by an hour to allow testing on the
rebase to continue while we find the cause of the increase in
time to upgrade the cluster.