openshift/os: increase requested resources
The `ostree` operations that happen as part of the `cosa-build` image
build are incredibly memory hungry during the `ostree commit` and
`ostree container` operations. They are moving upwards of 2G of data
into memory and onto disk and vice-versa.

The original resource requests were insufficient, causing the CI jobs
to be incredibly slow and sometimes even timing out completely. It's
been observed that the `ostree container encapsulate` operation ends
up requesting nearly 6Gi of memory.

This bumps both the memory requests and the CPU requests for the image
builds. It should give the jobs some healthy head room to perform the
operations at a reasonable pace.