Trigger assisted-ipv6 job only when asked
``e2e-metal-assisted-ipv6`` is not that valuable nowadays that we have
disconnected and dual-stack jobs implemented. That job uses a proxy
server and simulates installation of a ipv6 environment using that
ipv4-ipv6 proxy, all that under a cloud-like environment. On cloud users
will most likely do a flow of dual-stack (as cloud disables support for
IPv6) and for an IPv6-only environment do a flow like ZTP-disconnected

This change aims to not initiate job unless specifically asked by the
user (with ``/test e2e-metal-assisted-ipv6``. Reason for keeping this
job and the ability to run it is because it's good for validating proxy
usage (which isn't covered by any flow I think)