openshift/os: Copy all cluster jobs from MCO
Recently a lot of useful work landed in the CI for openshift/os,
in particular bits which actually launch a cluster, which closes
a huge gap.

However, the (previously named) `test-in-cluster` was just testing
that a cluster installs in GCP, and then we run a custom test
do do a build+rebase in cluster.

While that's a useful test, we actually want to be able to run
all the regular openshift e2e tests as well, for things like RHEL
rebases.

Rename `test-in-cluster` → `test-boot-derived-image` for clarity,
and while we're here make it `optional: true` because it's actually
broken right now since https://github.com/coreos/coreos-layering-examples/pull/21
landed - we need to change the test to synthesize a `rhel.repo`.