Increase baremetalds-packet-check number of seconds for Assisted
For assisted-installer related jobs, due to the hub-spoke nature, we
usually have to run more than one OpenShift installation as part of the
tests. This means packet machines in assisted jobs are needed for longer
and so the threshold of when a server is considered "leaked" should be
increased from 4 to 5 hours.

This commit makes it so that the baremetalds-packet-check step is
configurable so the of time before a server is considered a leak can be
set differently for each packet check job. The assisted jobs will now
use 5 hours for that step.

Also deleted assisted-packet-check and sno-packet-check from the
release-ocm-2.5 branch config because it's enough to just have it
on the master branch config, those machines are shared by all branches
anyway so cleaning leaks on one branch cleans for the others as well.