Add IMAGE_BUILDER env var
Add the IMAGE_BUILDER env var to support older builds which did not
support the SKIP_IMAGE_BUILD env var. So we can workarround it by
setting IMAGE_BUILDER to "echo" which will simply output the command. On
the other hand we need to set a default value for this, which we set to
podman in our case (even we use the builder from CI), because otherwise
we would run in issues as older releases make use of something like
"$IMAGE_BUILDER build . -f Dockerfile ...".

Signed-off-by: Christoph Stäbler <cstabler@redhat.com>