Permit TRT to approve/override for reverts TRT is responsible for the health of the CI signal for the release in main. If they detect a statistical increase in failure rates, they will try to find the PR responsible and revert it.
In order to quickly restore the CI signal, they need
to be able to do this without the repo OWNERS
participation if those OWNERS are not available
synchronously.

This change provides the team permission to override
tests across the openshift org and the ability
to apply the approved label. This can be used to
force reverts to merge (a separate TRT member must
supply an lgtm).