assisted-service-build copy the sources from src
These 2 lines:

https://github.com/openshift/assisted-service/blob/36a509d0075decdae0c0dc8a12200e0eed577088/Dockerfile.assisted-service-build#L27-L30

make running skipper commands very painful because they're extremely
slow (probably due to vendoring).

The reason they were added was to make the sources available as part of
the CI `assisted-service-build` image. But this can be achieved instead
using the `paths:` stanza to copy from the `src` image instead, as is
done in this commit.

After this is merged, I'll try to remove the offending lines from the
service and make sure the unit-tests (that use said image) still work.