Use deploy target for the compliance service e2e testing
We recently merged a new makefile target for deploying the compliance
service using kustomize.

  https://github.com/rhmdnd/compserv/commit/658c18acf215adeb904eb73bea41e3e34c6e78a3

Let's use this for the e2e-aws tests so that we exercise kustomize.
Eventually, we will update this to use a full e2e target that invokes
tests, but that will come at a later date.