MGMT-9378: Assisted Installer day-2 SNO periodics and presubmits
OCP 4.11 adds support for expanding single node clusters with additional
workers, so Assisted installer has been updated to support it as well.

This commit makes it so that we'd have a periodic test to make sure that
it works when using Assisted Installer with its REST API. The test is
currently using the `minimal` `TEST_SUITE` instead of the full one
because currently the CI is using a 4.10 `openshift-tests` binary for the
conformance tests (running against a 4.11 cluster) and when doing that
one of the tests present in the full suite breaks. Created MGMT-10922 to
track this.

This commit also adds a presubmit job to the assisted-service with
that same test