Update some jobs to run more often.
We're presently trying to analyze impacts if we switch to ovn as the
default. In comparing a number of jobs with their sdn/ovn equivalents,
we found several that are run hundreds of times a week in one
configuration, and once a day in the other. This is an attempt to
slightly balance the scales, moving a few jobs to run 4x a day instead
of once.

Also ovirt is modified to run both 3x a day instead of one 2x a day and
one 4x a day. (same overall number of jobs)