openstack: Add a UPI workflow
This patch enables a step-registry-based workflow for testing UPI on
OpenStack, for OpenShift 4.12+.

The install and deprovision steps pull snippets from the UPI
documentation of the Installer, and execute them.

The new workflow is added as an Installer presubmit and as a periodic.
No change is made for template-based UPI jobs on 4.11 and down.