Remove the template test defintion from rhmdnd/compserv
This job was created when we added initial support for OpenShift CI to
the repository. Since then, we setup an additional job that deploys the
compliance service using kustomize, and we have separate tests invoked
using GitHub actions.

This commit removes the `test` job since it's not particularly useful at
this point.