move assisted-tools related steps to workflows
In preparation for MGMT-9583 (allowing to trigger some periodic jobs
on-demand using cluster-bot) this change creates the relevant workflows
so that we can take leverage of the workflow-launch command.

Another thing it does is move the tools related steps/workflows to be
under assisted instead of assisted-baremetal, as it has no use of an
actual baremetal machine.