Use N2 instead of N1 instance types for GCP
We are consistently failing GCP with messages like this:

```
level=error msg=Error: Error waiting for instance to create: The zone
'projects/XXXXXXXXXXXXXXXXXXXXXXXX/zones/us-central1-c' does not have
enough resources available to fulfill the request. '(resource
type:compute)'.
```

The n1 instance types we are trying to use are older, so perhaps there
are fewer? This updates the main gcp ipi conf step to try to use n2 type
instances.