ComplianceAsCode/content: Cloning the e2e test suite instead of go getting it
We have started using go 1.18 in the e2e test suite, which means that
fetching the ocp4e2e module fails. See https://go.dev/doc/go1.18 (search
for "go get") for more details.

To work around that, we simply clone the repo. As a follow up
enhancement, we should use a tag from the clone, but for now we need to
unblock our CI suite.