MGMT-10256: remove assisted jobs for selected versions
Since we have another validation going on with
https://github.com/openshift-assisted/assisted-installer-deployment/blob/05514c359b1554d05a0ef7f34b011b2181ff32bd/tools/bump_ocp_releases.py,
there's no need for us to validate assisted flow with every created
payload for those versions. It takes a lots a of of money to provision
all those Equinix machines, and we've yet to find any regression using
those kind of nightly jobs.