Sync single-node-e2e-test-commands.sh with upstream openshift-e2e-test-commands
The former was originally copied from the latter to make a few changes
(number of parallel tests and how long we wait for image streams).

Every once in a while I sync the changes from the latter back to the
former to make sure it's up to date and works as expected.

Was reminded of it by https://github.com/openshift/release/pull/30412