MGMT-10946: fix memory and disk usage of ZTP jobs
Going through our jobs and match the provisioned machines with the
actual requirements for their specs:
* Memory is set to 16GiB (worker nodes requirement is actually 8GiB, but
  we don't have a way separate it in dev-scripts' extra workers).
* Installation disk size is set to 100GB.

I also removed values that match the declared defaults of the ztp/capi
workflows, so it now contains less duplication and should be easier to
read. As part of it the default networking also changed to IPv4 as it's
the most frequently used one.