MGMT-11209: Switch from s3.xlarge to m3.large for assisted
Most of the s3.xlarge machine that we lease from equinix are located in sp4 (Brazil) which costs 4.43$/h. This is likely due to a lack of availability of these kind of machines in other datacenters.

The m3.large is 5% more expensive in US (3.1$/h vs. 2.95$/h for the s3.xlarge), but has more availability globally.

The assumption is that even though the m3.large is more expensive, we will lease the machines more often in locations where the cost is cheap. Therefore, save money.