Drop core cloud serial jobs interval to 4h.
etcd team needs quick validation of a feature, we are temporarily
changing the interval on these jobs to each run as often as possible. This
change is made for aws-ovn nightly, aws-sdn ci, gcp-sdn ci, and
azure-sdn ci. aws-sdn nightly is left alone as aws-sdn is covered in the
ci stream.

This change will need to be reverted in the coming weeks.

testplatform has informed there is no way to run parallel periodics,
they'll just wait for the previous one to finish, thus I went with 4h
not the 1h we had planned.