os: Fix periodic extensions to pull from just-built image
I think what's going on with https://github.com/openshift/os/issues/908
is that the extensions build is trying to pull the *current*
RHCOS (which is broken) instead of the one we just built in the job.

And since the current RHCOS will only be fixed when this job
completes in total, we're in deadlock.