Modify golangci-lint job to run in a container (#30807)
* Add golangci-lint job to run in a container

Existing golangci job install golanci-lint binary which is failing
after receiving 429 (ratelimit) from github.[1]

If we run golangci-lint in container we don't need to install the
binary from github and can avoid the ratelimit issue completely.

With this commit, Adding a new job which run golangci-lint in
golangci-lint container.

[1] https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_release/30638/rehearse-30638-pull-ci-openstack-k8s-operators-mariadb-operator-master-golangci/1551583824574418944/build-log.txt

* Remove existing golangci jobs

Existing golangci job install golanci-lint binary which is failing
after receiving 429 (ratelimit) from github[1]. Removing this job
as we will use a job that will run the  `golanci-lint` inside
`golanci-lint` container.

[1] https://storage.googleapis.com/origin-ci-test/pr-logs/pull/openshift_release/30638/rehearse-30638-pull-ci-openstack-k8s-operators-mariadb-operator-master-golangci/1551583824574418944/build-log.txt

* Rename golangci-lint job as golangci

Renaming golangci-lint job as golangci as this was
the original name of golangci job.